Refactor: Extract AppBar and Drawer into separate components
- Extracted the AppBar JSX from App.jsx into a new MainAppBar.jsx component.
- Extracted the Drawer (Sidebar) JSX from App.jsx into a new Sidebar.jsx component.
- Refactored App.jsx to use the new components, passing props down accordingly.
- This change cleans up the main App component, making it easier to read and maintain.
